Tidied up and commented on the Activision stack implementation.
