+++ Machines/Atari2600/Cartridges/Cartridge.hpp
@@ -9,52 +9,10 @@
 #ifndef Atari2600_Cartridge_hpp
 #define Atari2600_Cartridge_hpp
 
-#include "../../Processors/6502/CPU6502.hpp"
+#include "../../../Processors/6502/CPU6502.hpp"
+#include "../Bus.hpp"
 
 namespace Atari2600 {
-
-class Bus {
-	public:
-		Bus() :
-			tia_input_value_{0xff, 0xff},
-			cycles_since_speaker_update_(0),
-			cycles_since_video_update_(0),
-			cycles_since_6532_update_(0) {}
-
-		virtual void run_for_cycles(int number_of_cycles) = 0;
-		virtual void set_reset_line(bool state) = 0;
-
-		// the RIOT, TIA and speaker
-		PIA mos6532_;
-		std::shared_ptr<TIA> tia_;
-		std::shared_ptr<Speaker> speaker_;
-
-		// joystick state
-		uint8_t tia_input_value_[2];
-
-	protected:
-		// speaker backlog accumlation counter
-		unsigned int cycles_since_speaker_update_;
-		inline void update_audio() {
-			unsigned int audio_cycles = cycles_since_speaker_update_ / 114;
-			cycles_since_speaker_update_ %= 114;
-			speaker_->run_for_cycles(audio_cycles);
-		}
-
-		// video backlog accumulation counter
-		unsigned int cycles_since_video_update_;
-		inline void update_video() {
-			tia_->run_for_cycles((int)cycles_since_video_update_);
-			cycles_since_video_update_ = 0;
-		}
-
-		// RIOT backlog accumulation counter
-		unsigned int cycles_since_6532_update_;
-		inline void update_6532() {
-			mos6532_.run_for_cycles(cycles_since_6532_update_);
-			cycles_since_6532_update_ = 0;
-		}
-};
 
 template<class T> class Cartridge:
 	public CPU6502::Processor<Cartridge<T>>,
