+++ Machines/Atari2600/Cartridges/CartridgeActivisionStack.hpp
@@ -22,6 +22,9 @@ class CartridgeActivisionStack: public Cartridge<CartridgeActivisionStack> {
 		void perform_bus_operation(CPU6502::BusOperation operation, uint16_t address, uint8_t *value) {
 			if(!(address & 0x1000)) return;
 
+			// This is a bit of a hack; a real cartridge can't see either the sync or read lines, and can't see
+			// address line 13. Instead it looks for a pattern in recent address accesses that would imply an
+			// RST or JSR.
 			if(operation == CPU6502::BusOperation::ReadOpcode && (last_opcode_ == 0x20 || last_opcode_ == 0x60)) {
 				if(address & 0x2000) {
 					rom_ptr_ = rom_.data();
