+++ Machines/Atari2600/Cartridges/CartridgeMegaBoy.hpp
